winkernel: move more vars to new mods
- moved loadbalancer info and flags vars to hns_util
- loadBalancerIngressInfo (used by service and hns) and externalIP to util
